+++ packages/grid/_modules_/grid/components/index.ts
@@ -8,7 +8,7 @@ export * from './toolbar';
 
 export * from './GridApiContext';
 export * from './GridAutoSizer';
-export * from './GridCell';
+export * from './cell/GridCell';
 export * from './GridCheckboxRenderer';
 export * from './GridFooter';
 export * from './GridHeader';
@@ -16,7 +16,7 @@ export * from './GridLoadingOverlay';
 export * from './GridNoRowsOverlay';
 export * from './GridPagination';
 export * from './GridRenderingZone';
-export * from './GridRowCells';
+export * from './cell/GridRowCells';
 export * from './GridRowCount';
 export * from './GridRow';
 export * from './GridSelectedRowCount';
@@ -24,4 +24,4 @@ export * from './GridStickyContainer';
 export * from './GridViewport';
 export * from './Watermark';
 export * from './GridScrollArea';
-export * from './GridEmptyCell';
+export * from './cell/GridEmptyCell';
